requirements/base.txt
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
python-dotenv==1.0.1
django_debug_toolbar==4.2.0
algoliasearch-django==3.0.0
djangorestframework-simplejwt==5.3.1
celery==5.3.6